Add files via upload
I initially tested data of different sizes (50k, 100k, 200k, 500k), and simply counted the number of comments and word frequencies on them. For all sizes of data sets, I kept the number of CPU cores available to each Executor unchanged (I set it to 2), and recorded the computing time with different numbers of Executors.